Allow running the process as a non root user.
I've chowned the /app directory in the container to node user.  This allows the process to access the private key.

However! I've not set the USER to node in the Dockerfile because then the process would be unable to bind to port 80 internally (or anything below 1024).
I didn't want to force it either and make it a breaking change, because many existing setups will be mapping 8080 externally to 80/443 internally, for example.

The workaround is therefore on the user (sorry), they will need to set the environment variable to a higher number, or use the --sysctl net.ipv4.ip_unprivileged_port_start=0 flag

Reference: https://github.com/moby/moby/issues/8460#issuecomment-312459310

Issue #14